Prise en compte des commentaires
- Garde un seul composant Bouton.
- Corrige les noms de fonction, dossier, fichier.
- Ajoute des {} dans les paramètres des fonctions.
- Supprime les fichiers inutiles.
- Range les dossiers dans les bonnes catégories (molécules ou atoms)
