+++ src/atoms/ligne/ligne.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function Ligne() {
+    return (
+        <div className="ligne">
+            <p>
+                Ligne
+            </p>
+        </div>
+    )
+}
