+++ src/atoms/pub/pub.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Pub() {
+    return (
+        <div className="pub">
+
+        </div>
+    )
+}
