+++ src/atoms/section/section.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function Section(titre) {
+    return (
+        <div className="section">
+            <p>
+                Section
+            </p>
+        </div>
+    )
+}
