+++ src/atoms/titre/titre.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function Titre() {
+    return (
+        <div className="titre">
+            <p>
+                Titre
+            </p>
+        </div>
+    )
+}
