+++ src/molecules/article/article.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Article({ src, categorie, titre, resumeArticle, tempsDeLecture, titreArticle, phraseDaccroche }) {
+    return (
+        <div className="article">
+            <p> Article </p>
+        </div>
+    )
+}
