+++ src/molecules/categorie/categorie.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function Categorie({ titre }) {
+    return (
+        <div className="categorie">
+            <p>
+                Catégorie
+            </p>
+        </div>
+    )
+}
