+++ src/molecules/categories-les-plus-lus/categories-les-plus-lus.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function CategoriesLesPlusLus() {
+    return (
+        <div className="categories-les-plus-lus">
+            <p>
+                Catégories les plus lus !
+            </p>
+        </div>
+    )
+}
