+++ src/molecules/fiche-resume-film/fiche-resume-film.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function FicheResumeFilm({ children, type, titre, auteur }) {
+    return (
+        <div>
+            <span className='type_film'></span>
+            <span className='titre_film'></span>
+            <span className='auteur_film'></span>
+        </div>
+    )
+}
