+++ src/molecules/header/header.jsx
@@ -1,11 +1,10 @@
 import React from 'react'
+import { PoweroffOutlined } from '@ant-design/icons';
 
 import "./header.css"
 
-import MenuHamburger from '../../atoms/menu-hamburger/menuHamburger'
-import TitreSite from '../../atoms/titre-site/titreSite'
-import BoutonLogin from '../../atoms/bouton-login/boutonLogin'
-import BoutonAbonnement from '../../atoms/bouton-abonnement/boutonAbonnement'
+import MenuHamburger from '../menu-hamburger/menuHamburger'
+import TitreSite from '../titre-site/titreSite'
 import Bouton from '../bouton/bouton'
 
 
@@ -14,9 +13,9 @@ export default function Header() {
         <div className="header">
             <MenuHamburger />
             <TitreSite />
-            <BoutonLogin />
-            <BoutonAbonnement />
-            <Bouton iconeDeGauche={<IconeConnexion />}>
+            <Bouton titre="login" />
+            <Bouton titre="abonnement" />
+            <Bouton iconeDeGauche={<PoweroffOutlined />}>
                 Se connecter
             </Bouton>
             <Bouton graisse="bold" backgroundColor="yellow">
