+++ src/molecules/mentions-legales/mentionsLegales.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function MentionsLegales({ titre }) {
+    return (
+        <div className="mentions-legales">
+            <p>
+                Mentions Légales
+            </p>
+        </div>
+    )
+}
