+++ src/molecules/pub/pub.jsx
@@ -5,7 +5,7 @@ import Pub from './pub.css'
 export default function Pub() {
     return (
         <div className="pub">
-            <Pub />
+            <p> Pub </p>
         </div>
     )
 }
