+++ src/molecules/section/section.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
 
-export default function Texte() {
+export default function Section() {
     return (
-        <div className="texte">
+        <div className="section">
             <p>
-                Texte
+                Section
             </p>
         </div>
     )
