+++ src/molecules/sous-categorie/sousCategorie.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function SousCategorie({ titre, text_decoration, text_decoration_color }) {
+    return (
+        <div className="sous-categorie">
+            <p>
+                Sous-catégorie
+            </p>
+        </div>
+    )
+}
