+++ src/molecules/titre-site/titreSite.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function TitreSite() {
+    return (
+        <div className="titre-site">
+            <p>
+                Titre Site
+            </p>
+        </div>
+    )
+}
