+++ src/molecules/top-categories-les-plus-lus/topCategoriesLesPlusLus.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function TopCategoriesLesPlusLus({ classement, titreCategorie, TitreArticle }) {
+    return (
+        <div className="top-categories-les-plus-lus">
+            <p>
+                Top catégories les plus lus
+            </p>
+        </div>
+    )
+}
